+++ src/components-deprecated/home/index.ts
@@ -1,3 +1,2 @@
-export { default as HomeHeader } from './HomeHeader';
 export { default as RoomCreateButton } from './RoomCreateButton';
 export { default as RoomList } from './RoomList';
